+++ src/Hakyll/Check.hs
@@ -94,7 +94,8 @@ runChecker checker config verbosity check' = do
 checkDestination :: Checker ()
 checkDestination = do
     config <- checkerConfig <$> ask
-    files  <- liftIO $ getRecursiveContents (destinationDirectory config)
+    files  <- liftIO $
+        getRecursiveContents (const False) (destinationDirectory config)
 
     let htmls =
             [ destinationDirectory config </> file