+++ src/Hakyll/Web/Pandoc/FileType.hs
@@ -8,7 +8,7 @@ module Hakyll.Web.Pandoc.FileType
 
 
 --------------------------------------------------------------------------------
-import           System.FilePath        (takeExtension)
+import           System.FilePath        (splitExtension)
 
 
 --------------------------------------------------------------------------------
@@ -36,27 +36,31 @@ data FileType
 --------------------------------------------------------------------------------
 -- | Get the file type for a certain file. The type is determined by extension.
 fileType :: FilePath -> FileType
-fileType = fileType' . takeExtension
+fileType = uncurry fileType' . splitExtension
   where
-    fileType' ".css"      = Css
-    fileType' ".htm"      = Html
-    fileType' ".html"     = Html
-    fileType' ".lhs"      = LiterateHaskell Markdown
-    fileType' ".markdown" = Markdown
-    fileType' ".md"       = Markdown
-    fileType' ".mdn"      = Markdown
-    fileType' ".mdown"    = Markdown
-    fileType' ".mdwn"     = Markdown
-    fileType' ".mkd"      = Markdown
-    fileType' ".mkdwn"    = Markdown
-    fileType' ".org"      = OrgMode
-    fileType' ".page"     = Markdown
-    fileType' ".rst"      = Rst
-    fileType' ".tex"      = LaTeX
-    fileType' ".text"     = PlainText
-    fileType' ".textile"  = Textile
-    fileType' ".txt"      = PlainText
-    fileType' _           = Binary  -- Treat unknown files as binary
+    fileType' _ ".css"      = Css
+    fileType' _ ".htm"      = Html
+    fileType' _ ".html"     = Html
+    fileType' f ".lhs"      = LiterateHaskell $ case fileType f of
+        -- If no extension is given, default to Markdown + LiterateHaskell
+        Binary -> Markdown
+        -- Otherwise, LaTeX + LiterateHaskell or whatever the user specified
+        x      -> x
+    fileType' _ ".markdown" = Markdown
+    fileType' _ ".md"       = Markdown
+    fileType' _ ".mdn"      = Markdown
+    fileType' _ ".mdown"    = Markdown
+    fileType' _ ".mdwn"     = Markdown
+    fileType' _ ".mkd"      = Markdown
+    fileType' _ ".mkdwn"    = Markdown
+    fileType' _ ".org"      = OrgMode
+    fileType' _ ".page"     = Markdown
+    fileType' _ ".rst"      = Rst
+    fileType' _ ".tex"      = LaTeX
+    fileType' _ ".text"     = PlainText
+    fileType' _ ".textile"  = Textile
+    fileType' _ ".txt"      = PlainText
+    fileType' _ _           = Binary  -- Treat unknown files as binary
 
 
 --------------------------------------------------------------------------------