+++ web/releases.markdown
@@ -4,6 +4,18 @@ title: Releases
 
 # Releases
 
+## Hakyll 4.1.4.0
+
+*January 26, 2013*
+
+- Export the flexible `renderTags` function
+
+## Hakyll 4.1.3.0
+
+*January 26, 2013*
+
+- Export the constructor of the `Tags` datatype
+
 ## Hakyll 4.1.2.0
 
 *January 20, 2013*